requirements.txt
accelerate==1.1.1
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.11.8
aiosignal==1.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.6.2.post1
attrs==24.2.0
bidict==0.23.1
botorch==0.12.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
diffusers==0.31.0
docutils==0.21.2
fastapi==0.115.5
filelock==3.16.1
fonttools==4.55.0
frozenlist==1.5.0
fsspec==2024.10.0
gpytorch==1.13
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.0
huggingface-hub==0.26.3
hydra-core==1.3.2
idna==3.10
ifaddr==0.2.0
importlib_metadata==8.5.0
itsdangerous==2.2.0
jaxtyping==0.2.19
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
linear-operator==0.5.3
markdown2==2.5.1
MarkupSafe==3.0.2
matplotlib==3.9.2
mpmath==1.3.0
multidict==6.1.0
multipledispatch==1.0.0
networkx==3.4.2
nicegui==2.7.0
numpy==2.1.3
omegaconf==2.3.0
opt_einsum==3.4.0
orjson==3.10.12
packaging==24.2
pillow==11.0.0
propcache==0.2.0
pscript==0.7.7
psutil==6.1.0
pydantic==2.10.2
pydantic_core==2.27.1
Pygments==2.18.0
pyparsing==3.2.0
pyro-api==0.1.2
pyro-ppl==1.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.10.1
python-multipart==0.0.18
python-socketio==5.11.4
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
simple-websocket==1.1.0
six==1.16.0
sniffio==1.3.1
starlette==0.41.3
sympy==1.13.1
threadpoolctl==3.5.0
tokenizers==0.20.3
torch==2.5.1
tqdm==4.67.1
transformers==4.46.3
typeguard==4.4.1
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.1
vbuild==0.8.2
watchfiles==1.0.0
websockets==14.1
wsproto==1.2.0
yarl==1.18.0
zipp==3.21.0
